+++ include/modules/ipc.hpp
@@ -35,9 +35,17 @@ namespace modules {
     static constexpr auto TYPE = "custom/ipc";
 
     static constexpr auto EVENT_SEND = "send";
+    static constexpr auto EVENT_HOOK = "hook";
+    static constexpr auto EVENT_NEXT = "next";
+    static constexpr auto EVENT_PREV = "prev";
+    static constexpr auto EVENT_RESET = "reset";
 
    protected:
     void action_send(const string& data);
+    void action_hook(const string& data);
+    void action_next();
+    void action_prev();
+    void action_reset();
 
    private:
     static constexpr const char* TAG_OUTPUT{"<output>"};
@@ -45,6 +53,8 @@ namespace modules {
     map<mousebtn, string> m_actions;
     string m_output;
     size_t m_initial;
+    size_t m_current_hook;
+    void exec_hook();
   };
 }  // namespace modules