+++ src/modules/ipc.cpp
@@ -14,8 +14,12 @@ namespace modules {
    * Load user-defined ipc hooks and
    * create formatting tags
    */
-  ipc_module::ipc_module(const bar_settings& bar, string name_) : module<ipc_module>(bar, move(name_)) {
+  ipc_module::ipc_module(const bar_settings& bar, string name_) : module<ipc_module>(bar, move(name_)), m_initial(0) {
     m_router->register_action_with_data(EVENT_SEND, [this](const std::string& data) { action_send(data); });
+    m_router->register_action_with_data(EVENT_HOOK, [this](const std::string& data) { action_hook(data); });
+    m_router->register_action(EVENT_NEXT, [this]() { action_next(); });
+    m_router->register_action(EVENT_PREV, [this]() { action_prev(); });
+    m_router->register_action(EVENT_RESET, [this]() { action_reset(); });
 
     size_t index = 0;
 
@@ -25,8 +29,16 @@ namespace modules {
 
     m_log.info("%s: Loaded %d hooks", name(), m_hooks.size());
 
-    if ((m_initial = m_conf.get(name(), "initial", 0_z)) && m_initial > m_hooks.size()) {
-      throw module_error("Initial hook out of bounds (defined: " + to_string(m_hooks.size()) + ")");
+    m_initial = m_conf.get(name(), "initial", 0_z);
+    if (m_conf.has(name(), "initial") && m_initial != 0) {
+      if (m_initial <= m_hooks.size()) {
+        m_current_hook = m_initial - 1;
+      } else {
+        throw module_error("Initial hook out of bounds '" + to_string(m_initial) + "'. Defined hooks goes from 1 to " +
+                           to_string(m_hooks.size()) + ")");
+      }
+    } else {
+      m_current_hook = m_hooks.size();
     }
 
     // clang-format off
@@ -62,11 +74,9 @@ namespace modules {
       broadcast();
     });
 
-    if (m_initial) {
+    if (m_initial != 0) {
       // TODO do this in a thread.
-      auto command = command_util::make_command<output_policy::REDIRECTED>(m_hooks.at(m_initial - 1)->command);
-      command->exec(false);
-      command->tail([this](string line) { m_output = line; });
+      exec_hook();
     }
   }
 
@@ -107,30 +117,82 @@ namespace modules {
    * execute its command
    */
   void ipc_module::on_message(const string& message) {
-    for (auto&& hook : m_hooks) {
-      if (hook->payload != message) {
-        continue;
+    for (size_t i = 0; i < m_hooks.size(); i++) {
+      const auto& hook = m_hooks[i];
+      if (hook->payload == message) {
+        m_log.info("%s: Found matching hook (%s)", name(), hook->payload);
+        m_current_hook = i;
+        this->exec_hook();
+        break;
       }
-
-      m_log.info("%s: Found matching hook (%s)", name(), hook->payload);
-
-      try {
-        // Clear the output in case the command produces no output
-        m_output.clear();
-        auto command = command_util::make_command<output_policy::REDIRECTED>(hook->command);
-        command->exec(false);
-        command->tail([this](string line) { m_output = line; });
-      } catch (const exception& err) {
-        m_log.err("%s: Failed to execute hook command (err: %s)", err.what());
-        m_output.clear();
-      }
-
-      broadcast();
     }
   }
 
   void ipc_module::action_send(const string& data) {
     m_output = data;
+    broadcast();
+  }
+
+  void ipc_module::action_hook(const string& data) {
+    try {
+      int hook = std::stoi(data);
+
+      if (hook < 0 || (size_t)hook >= m_hooks.size()) {
+        m_log.err("%s: Hook action received with an out of bounds hook '%s'. Defined hooks goes from 0 to %zu.", name(),
+            data, m_hooks.size() - 1);
+      } else {
+        m_current_hook = hook;
+        this->exec_hook();
+      }
+    } catch (const std::invalid_argument& err) {
+      m_log.err(
+          "%s: Hook action received '%s' cannot be converted to a valid hook index. Defined hooks goes from 0 to %zu.",
+          name(), data, m_hooks.size() - 1);
+    }
+  }
+
+  void ipc_module::action_next() {
+    // this is the case where initial is not defined on first 'next'
+    if (m_current_hook == m_hooks.size()) {
+      m_current_hook = 0;
+    } else {
+      m_current_hook = (m_current_hook + 1) % m_hooks.size();
+    }
+    this->exec_hook();
+  }
+
+  void ipc_module::action_prev() {
+    if (m_current_hook == 0) {
+      m_current_hook = m_hooks.size();
+    }
+    m_current_hook--;
+    this->exec_hook();
+  }
+
+  void ipc_module::action_reset() {
+    if (m_initial != 0) {
+      m_current_hook = m_initial - 1;
+      this->exec_hook();
+    } else {
+      m_current_hook = m_hooks.size();
+      m_output.clear();
+      broadcast();
+    }
+  }
+
+  void ipc_module::exec_hook() {
+    // Clear the output in case the command produces no output
+    m_output.clear();
+
+    try {
+      auto command = command_util::make_command<output_policy::REDIRECTED>(m_hooks[m_current_hook]->command);
+      command->exec(false);
+      command->tail([this](string line) { m_output = line; });
+    } catch (const exception& err) {
+      m_log.err("%s: Failed to execute hook command (err: %s)", name(), err.what());
+      m_output.clear();
+    }
+
     broadcast();
   }
 }  // namespace modules